+++ src/main/gen/com/github/nthykier/debpkg/aptlist/psi/AptListOption.java
@@ -1,0 +1,16 @@
+// This is a generated file. Not intended for manual editing.
+package com.github.nthykier.debpkg.aptlist.psi;
+
+import java.util.List;
+import org.jetbrains.annotations.*;
+import com.intellij.psi.PsiElement;
+
+public interface AptListOption extends PsiElement {
+
+  @NotNull
+  AptListOptionKey getOptionKey();
+
+  @NotNull
+  List<AptListOptionValue> getOptionValueList();
+
+}