+++ src/main/gen/com/github/nthykier/debpkg/aptlist/psi/AptListVisitor.java
@@ -1,0 +1,46 @@
+// This is a generated file. Not intended for manual editing.
+package com.github.nthykier.debpkg.aptlist.psi;
+
+import org.jetbrains.annotations.*;
+import com.intellij.psi.PsiElementVisitor;
+import com.intellij.psi.PsiElement;
+
+public class AptListVisitor extends PsiElementVisitor {
+
+  public void visitAptSourcesLine(@NotNull AptListAptSourcesLine o) {
+    visitPsiElement(o);
+  }
+
+  public void visitComponent(@NotNull AptListComponent o) {
+    visitPsiElement(o);
+  }
+
+  public void visitOption(@NotNull AptListOption o) {
+    visitPsiElement(o);
+  }
+
+  public void visitOptionKey(@NotNull AptListOptionKey o) {
+    visitPsiElement(o);
+  }
+
+  public void visitOptionValue(@NotNull AptListOptionValue o) {
+    visitPsiElement(o);
+  }
+
+  public void visitSuite(@NotNull AptListSuite o) {
+    visitPsiElement(o);
+  }
+
+  public void visitType(@NotNull AptListType o) {
+    visitPsiElement(o);
+  }
+
+  public void visitUri(@NotNull AptListUri o) {
+    visitPsiElement(o);
+  }
+
+  public void visitPsiElement(@NotNull PsiElement o) {
+    visitElement(o);
+  }
+
+}