+++ src/main/java/com/github/nthykier/debpkg/aptlist/AptListLanguage.java
@@ -1,0 +1,11 @@
+package com.github.nthykier.debpkg.aptlist;
+
+import com.intellij.lang.Language;
+
+public class AptListLanguage extends Language {
+    public static final AptListLanguage INSTANCE = new AptListLanguage();
+
+    private AptListLanguage() {
+        super("DebianAptSourcesList");
+    }
+}