+++ CHANGELOG.md
@@ -13,6 +13,7 @@
 - Add support for IDEA 2023.1
 - Add support for apt deb822-based sources files.
 - Add `needs-sudo` as known restriction for `debian/tests/control`.  (Closes: #112)
+- Add basic syntax highlighting of APT's `.list` sources file. (Closes: #114)
 
 ### Fixed
 - Avoid assertion error on creating a new (blank) `debian/control` and the first field typed is *not* `Source:`.  (Closes: #107)